Extract common code fix code to BaseConditionConstraintCodeFix
