+++ src/nunit.analyzers/ConstraintUsage/EqualConstraintUsageCodeFix.cs
@@ -1,65 +1,24 @@
 using System.Collections.Immutable;
 using System.Composition;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CodeActions;
 using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using NUnit.Analyzers.Constants;
-using NUnit.Analyzers.Extensions;
-using static NUnit.Analyzers.Constants.NunitFrameworkConstants;
 
 namespace NUnit.Analyzers.ConstraintUsage
 {
     [ExportCodeFixProvider(LanguageNames.CSharp)]
     [Shared]
-    public class EqualConstraintUsageCodeFix : CodeFixProvider
+    public class EqualConstraintUsageCodeFix : BaseConditionConstraintCodeFix
     {
         public override ImmutableArray<string> FixableDiagnosticIds { get; } = ImmutableArray.Create(AnalyzerIdentifiers.EqualConstraintUsage);
 
-        public sealed override FixAllProvider GetFixAllProvider()
+        protected override (ExpressionSyntax actual, ExpressionSyntax constraintExpression) GetActualAndConstraintExpression(ExpressionSyntax conditionNode, string suggestedConstraintString)
         {
-            return WellKnownFixAllProviders.BatchFixer;
-        }
-
-        public override async Task RegisterCodeFixesAsync(CodeFixContext context)
-        {
-            var suggestedConstraintString = context.Diagnostics[0].Properties[EqualConstraintUsageAnalyzer.SuggestedConstraintString];
-
-            var description = string.Format(CodeFixConstants.UseConstraintDescriptionFormat, suggestedConstraintString);
-
-            var root = await context.Document.GetSyntaxRootAsync(context.CancellationToken).ConfigureAwait(false);
-            var semanticModel = await context.Document.GetSemanticModelAsync(context.CancellationToken).ConfigureAwait(false);
-
-            context.CancellationToken.ThrowIfCancellationRequested();
-            var conditionNode = (root.FindNode(context.Span) as ArgumentSyntax)?.Expression;
-
-            if (conditionNode == null)
-                return;
-
             var (actual, expected) = GetActualExpected(conditionNode);
-
-            var assertNode = conditionNode.Ancestors()
-                .OfType<InvocationExpressionSyntax>()
-                .FirstOrDefault();
-
-            var assertMethod = semanticModel.GetSymbolInfo(assertNode).Symbol as IMethodSymbol;
-
-            if (actual == null || expected == null || assertNode == null || assertMethod == null)
-                return;
-
-            var newAssertNode = UpdateAssertNode(assertNode, assertMethod, actual, expected, suggestedConstraintString);
-            var newRoot = root.ReplaceNode(assertNode, newAssertNode);
-
-            var codeAction = CodeAction.Create(
-                description,
-                _ => Task.FromResult(context.Document.WithSyntaxRoot(newRoot)),
-                description);
-
-            context.RegisterCodeFix(codeAction, context.Diagnostics);
-
+            var constraintExpression = GetConstraintExpression(suggestedConstraintString, expected);
+            return (actual, constraintExpression);
         }
 
         private static (ExpressionSyntax actual, ExpressionSyntax expected) GetActualExpected(SyntaxNode conditionNode)
@@ -97,37 +56,6 @@ namespace NUnit.Analyzers.ConstraintUsage
             }
 
             return (null, null);
-        }
-
-        private static InvocationExpressionSyntax UpdateAssertNode(InvocationExpressionSyntax assertNode, IMethodSymbol assertMethod,
-            ExpressionSyntax actual, ExpressionSyntax expected,
-            string constraintString)
-        {
-            // Replace Assert method to Assert.That
-            var newExpression = ((MemberAccessExpressionSyntax)assertNode.Expression)
-                .WithName(SyntaxFactory.IdentifierName(NameOfAssertThat));
-
-            // Replace arguments
-            var hasConstraint = assertNode.GetArgumentExpression(assertMethod, NameOfExpressionParameter) != null;
-
-            var remainingArguments = hasConstraint
-                ? assertNode.ArgumentList.Arguments.Skip(2)
-                : assertNode.ArgumentList.Arguments.Skip(1);
-
-            var constraintExpression = SyntaxFactory.InvocationExpression(
-                SyntaxFactory.ParseExpression(constraintString),
-                SyntaxFactory.ArgumentList(SyntaxFactory.SingletonSeparatedList(SyntaxFactory.Argument(expected))));
-
-            var newArguments = new[] {
-                SyntaxFactory.Argument(actual),
-                SyntaxFactory.Argument(constraintExpression)
-            }.Union(remainingArguments);
-
-            var newArgumentsList = SyntaxFactory.ArgumentList(SyntaxFactory.SeparatedList(newArguments));
-
-            return assertNode
-                .WithExpression(newExpression)
-                .WithArgumentList(newArgumentsList);
         }
     }
 }
